webapi/webapi-iap-xwalk-tests/iap/Navigator_iap_init.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>WebAPI: IAP</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Zhong Qiu" href="mailto:zhongx.qiu@intel.com">
<link rel="help" href="http://www.w3.org/TR/IAP-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="options.js"></script>
<div id="log"></div>
<script>
  at1 = async_test("Check if the navigator.iap.init works well");
  at2 = async_test("Check if the second navigator.iap.init raise an InvalidStateError");

  init_promise = navigator.iap.init(options);

  at1.step(function() {
    init_promise.then(
      at1.step_func(function() {
        assert_not_equals(navigator.iap, null, "nagivator.iap is not null.");
        assert_true(init_promise instanceof Promise, "navigator.iap.init returns a promise");
        at1.done();

        at2.step(function() {
          navigator.iap.init(options).then(
            at2.step_func(function() {
              assert_unreached("Here should neve be reached!");
              at2.done();
            }),
            at2.step_func(function(error) {
              assert_equals(error.name, "InvalidStateError", "If the iap has already been initialized, return a Promise rejected with a DOMException named 'InvalidStateError'");
              at2.done();
            })
          );
        });
      }),

      at1.step_func(function(error) {
        assert_unreached("IAP initlialization failed");
        at2.done();
      })
    );
  });
</script>
